Move packages and update imports
Refactor package layout and update usages: moved RollbarManager from ti4.rollbar to ti4.logging, moved DataMigrationManager and MigrationHelper to ti4.game.persistence.migration, and moved AutoCompleteProvider into ti4.listeners. Updated imports across multiple classes (AsyncTI4DiscordBot, ButtonProcessor, CommandGameState, RunManualDataMigration, ExecutionHistoryManager, ContextMenuListener, ModalListener, SlashCommandListener, SelectionMenuProcessor, RequestContext, JdaService, ErrorLoggingFilter, etc.) and removed a now-unused import in AutoCompleteListener. This is a non-functional reorganization to better group logging and migration code and keep listener-related classes together.
