+++ src/main/java/ti4/AsyncTI4DiscordBot.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import ti4.logging.BotLogger;
-import ti4.rollbar.RollbarManager;
+import ti4.logging.RollbarManager;
 import ti4.settings.GlobalSettings;
 import ti4.settings.GlobalSettings.ImplementedSettings;
