+++ src/main/java/ti4/buttons/ButtonProcessor.java
@@ -30,8 +30,8 @@ import ti4.listeners.annotations.ButtonHandler;
 import ti4.listeners.context.ButtonContext;
 import ti4.logging.BotLogger;
 import ti4.logging.LogOrigin;
+import ti4.logging.RollbarManager;
 import ti4.message.MessageHelper;
-import ti4.rollbar.RollbarManager;
 import ti4.service.button.ReactionService;
 import ti4.service.game.GameNameService;
 import ti4.service.strategycard.PlayStrategyCardService;
