+++ src/main/java/ti4/commands/developer/RunManualDataMigration.java
@@ -8,11 +8,11 @@ import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import ti4.commands.Subcommand;
 import ti4.game.Game;
 import ti4.game.persistence.GameManager;
+import ti4.game.persistence.migration.DataMigrationManager;
 import ti4.helpers.Constants;
 import ti4.logging.BotLogger;
 import ti4.logging.LogOrigin;
 import ti4.message.MessageHelper;
-import ti4.migration.DataMigrationManager;
 
 class RunManualDataMigration extends Subcommand {
