+++ src/main/java/ti4/listeners/AutoCompleteListener.java
@@ -3,7 +3,6 @@ package ti4.listeners;
 import javax.annotation.Nonnull;
 import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import ti4.autocomplete.AutoCompleteProvider;
 import ti4.executors.ExecutorServiceManager;
 import ti4.spring.jda.JdaService;
