+++ src/main/java/ti4/listeners/AutoCompleteProvider.java
@@ -1,4 +1,4 @@
-package ti4.autocomplete;
+package ti4.listeners;
 
 import java.util.ArrayList;
 import java.util.Arrays;
