+++ src/main/java/ti4/listeners/ContextMenuListener.java
@@ -10,7 +10,7 @@ import ti4.commands.context.ContextCommand;
 import ti4.commands.context.ContextCommandManager;
 import ti4.executors.ExecutorServiceManager;
 import ti4.logging.BotLogger;
-import ti4.rollbar.RollbarManager;
+import ti4.logging.RollbarManager;
 import ti4.service.game.GameNameService;
 
 public class ContextMenuListener extends ListenerAdapter implements ListenerInterface {
