+++ src/main/java/ti4/spring/jda/JdaService.java
@@ -48,6 +48,7 @@ import ti4.cron.UploadStatsCron;
 import ti4.cron.WinningPathCron;
 import ti4.executors.ExecutorServiceManager;
 import ti4.game.persistence.GameManager;
+import ti4.game.persistence.migration.DataMigrationManager;
 import ti4.helpers.AliasHandler;
 import ti4.helpers.Constants;
 import ti4.helpers.Storage;
@@ -72,7 +73,6 @@ import ti4.listeners.UserJoinServerListener;
 import ti4.listeners.UserLeaveServerListener;
 import ti4.logging.BotLogger;
 import ti4.logging.LogBufferManager;
-import ti4.migration.DataMigrationManager;
 import ti4.selections.SelectionManager;
 import ti4.service.draft.SliceGenerationPipeline;
 import ti4.service.emoji.ApplicationEmojiService;
